+++ src/components/delete-button/index.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import Button from '../button';
-import iconDelete from './delete.svg';
+import Icon from '../icon';
 import './styles.scss';
 
 export default props => (
   <Button size="sm" className="mb-0" action="delete" {...props}>
-    <img src={iconDelete} alt="delete" />Delete
+    <Icon imgSrc="delete" alt="delete" />Delete
   </Button>
 );